Made get_free_memory_size() acquire the page stack lock

+++ kernel/page.h
@@ -67,7 +67,7 @@ err_t page_alloc_init(void);
 u64 page_alloc(void);
 u64 page_alloc_clear(void);
 void page_free(u64 page);
-u64 get_free_memory_size(void);
+size_t get_free_memory_size(void);
 err_t map_kernel_pages(u64 start, u64 length, bool write, bool execute);
 err_t map_user_pages(u64 start, u64 length, bool write, bool execute);
 void page_map_free_contents(u64 page_map_addr);
